LayoutTests/fast/events/onbeforeunload-focused-iframe-expected.txt

CONSOLE MESSAGE: beforeUnload
New window with beforeUnload handler.
You should see "beforeUnload" in the console log if you try to close the window.


--------
Frame: 'focusedFrame'
--------
This frame should get the focus.
